feat: share link background images, block manually setting images
database schema has changed

+++ src/routes/dash/shares/+page.server.ts
@@ -163,7 +163,9 @@ export const actions = {
     }
 
     // Delete the associated share
-    await db.delete(shares).where(eq(shares.id, locals.user.subdomainShareId));
+    if (locals.user.subdomainShareId) {
+      await db.delete(shares).where(eq(shares.id, locals.user.subdomainShareId));
+    }
 
     // Update user's subdomain to null
     const updated = await db
